Added TorqueEntity class
Allows the creation of objects whose vectors affect its rotation. Right
now VectorEntity and TorqueEntity are mutually exclusive, but they
could potentially be combined in the future if necessary. An
application of the class can be seen with the diamonds. When
instantiated, they sway back and forth. Movement isn’t as smooth as it
could be, will fix later.