Gravity update
Gravity takes into account platforms as well as the ground. Objects
still clip through the platforms and ground sometimes when they land on
them. I suspect this is because of the nature of move(), which sets the
new location of an object based on its xComp and yComp and as a result
only detects the ground after the object has moved and not during the
move.