Fixed Enemy bug
Enemies previously had collision detection issues with walls at certain
heights. Problem was solved by searching for walls at the four corners
of the image instead of just two points placed at the top and bottom.
The new issue is that isTouchingWall() uses this and lookForWall()
doesn’t, because when applied to the player in lookForWall(), it messes
with corner detections. Enemy is also still unable to turn around when
it reaches the edge of a platform.