Added Coordinate class
Trying out new form of collision detection in which we store
coordinates in arrays for each object/image that will be important in
collision detection (i.e. at corners, top, bottom, etc.) and the
methods in Force take these coordinates, decide which ones are to be
used for which type of detection, and applies collisions to them. Our
current system would probably have to be tinkered with to work for each
new image. Hopefully this system will allow us to avoid that.