Added lookFor methods
Methods with names such as isTouchingGround2() have been added and
given titles such as lookForGround2(). isTouchingRightWall() and
isTouchingLeftWall() have been added as well. Implementation of the new
collision system in the form of move2() has been unsuccessful so far;
the player can’t jump and the strange left force returns.