Bullet implemented in TestMan
The player (or rather TestMan) can now shoot bullets. The direction of
the bullet is controlled by the left and right arrow keys, so the
movement of the player has been changed to WASD. Ammo and reload time
have been implemented as well, and the bullets kill enemies and
disappear at walls. There’s currently an error where they almost always
don’t disappear when they reach the edge of the world.