Applied gravity to Diamond and Ammo
Wrote the gravity() command into the act methods for Diamonds and Ammo
so that when placed in the world, they always fall to the ground. To
make them continue to hover, I edited the dimensions of their pngs.
This guarantees they’re always at the same height, so we don’t have to
worry about how high we place them. This will also allow us to
randomize the ammo drops either when levels are instantiated or during
the gameplay, as we just have to spawn them at random points and
they’ll automatically land on platforms or the ground.